requirements.txt
streamlit
docxtpl
python-docx
pymupdf
Pillow
pandas